Fix end to end tests
