+++ build.sbt
@@ -20,9 +20,7 @@ lazy val commonSettings = Seq(
     Logging.scalaLogging,
     Logging.logBack,
     Logging.catsLogging,
-    Time.nScalaTime,
-    Testing.testFramework % "test",
-    Testing.testMockFramework % "test"
+    Time.nScalaTime
   ),
   dependencyOverrides ++= Dependencies.overrides,
   publish := {},
@@ -30,9 +28,19 @@ lazy val commonSettings = Seq(
   Test / fork := true
 )
 
+lazy val testingFramework = Seq(
+  libraryDependencies ++= List(
+    Testing.testFramework     % Test,
+    Testing.testMockFramework % Test,
+    Testing.testContainer     % Test,
+    Testing.simpleHttpClient  % Test,
+    Doobie.`doobie-scalatest` % Test
+  )
+)
+
 lazy val `domain` = (project in file("./domain"))
   .dependsOn(`kafka`, `test-support`)
-  .settings(commonSettings)
+  .settings(commonSettings, testingFramework)
   .settings(
     name := "domain",
     libraryDependencies ++= List(
@@ -61,7 +69,7 @@ lazy val `app` = (project in file("./app"))
 
 lazy val `end-to-end` = (project in file("./end-to-end"))
   .dependsOn(`test-support`, `domain`, `app`)
-  .settings(commonSettings)
+  .settings(commonSettings, testingFramework)
   .settings(Http4s.http4sAll)
   .settings(
     name := "end-to-end",
@@ -76,8 +84,8 @@ lazy val `end-to-end` = (project in file("./end-to-end"))
   )
 
 lazy val `web` = (project in file("./web"))
-  .dependsOn(`domain`, `test-support`)
-  .settings(commonSettings)
+  .dependsOn(`domain`, `web-domain`, `test-support`)
+  .settings(commonSettings, testingFramework)
   .settings(Http4s.http4sAll)
   .settings(
     name := "web",
@@ -93,15 +101,12 @@ lazy val `web` = (project in file("./web"))
       Cassandra.dataStax,
       Cassandra.dataStaxQueryBuilder,
       Doobie.`doobie-core`,
-      Doobie.`doobie-postgresql`,
-      Doobie.`doobie-scalatest`          % Test,
-      Testing.testContainer              % Test,
-      Testing.`testContainer-cassandra`  % Test,
-      Testing.`testContainer-postgresql` % Test
+      Doobie.`doobie-postgresql`
     )
   )
 
 lazy val `test-support` = (project in file("./test-support"))
+  .dependsOn(`web-domain`)
   .settings(commonSettings)
   .settings(Http4s.http4sAll)
   .settings(
@@ -117,9 +122,12 @@ lazy val `test-support` = (project in file("./test-support"))
       Doobie.`doobie-core`,
       Doobie.`doobie-postgresql`,
       Testing.`testContainer-kafka`,
-      Testing.simpleHttpClient,
       Testing.`testContainer-postgresql`,
-      Doobie.`doobie-scalatest` % "test",
+      Testing.testFramework,
+      Testing.testMockFramework,
+      Testing.testContainer,
+      Testing.simpleHttpClient,
+      Doobie.`doobie-scalatest`
     )
   )
 
@@ -129,7 +137,12 @@ lazy val `kafka` = (project in file("./kafka"))
     name := "kafka",
     libraryDependencies ++= List(
       Kafka.`fs2-kafka`,
-      Kafka.`fs2-kafka-vulcan`,
-      Testing.simpleHttpClient % Test
+      Kafka.`fs2-kafka-vulcan`
     )
+  )
+
+lazy val `web-domain` = (project in file("./web-domain"))
+  .settings(commonSettings)
+  .settings(
+    name := "web-domain"
   )
