--- domain/src/main/scala/com/kelvin/whoseturn/entity/TodoItemEntity.scala
@@ -1,19 +1,0 @@
-package com.kelvin.whoseturn.entity
-
-import java.sql.Timestamp
-
-import java.util.UUID
-
-case class TodoItemEntity(
-    id: UUID,
-    title: String,
-    createdBy: UUID,
-    createdOn: Timestamp,
-    lastUpdate: Timestamp,
-    description: String,
-    flagged: Boolean,
-    category: String,
-    priority: String,
-    location: String,
-    active: Boolean
-)
