+++ test-support/src/main/scala/com/kelvin/whoseturn/support/http/MockHttp.scala
@@ -1,0 +1,16 @@
+package com.kelvin.whoseturn.support.http
+
+import java.net.ServerSocket
+
+object MockHttp {
+
+  protected def getFreePort: Int = {
+    val socket = new ServerSocket(0)
+    socket.setReuseAddress(true)
+
+    val localPort = socket.getLocalPort
+    socket.close()
+
+    localPort
+  }
+}
