+++ web-domain/src/main/scala/com/kelvin/whoseturn/models/CreateTodoItemModel.scala
@@ -1,0 +1,11 @@
+package com.kelvin.whoseturn.models
+
+final case class CreateTodoItemModel(
+    title: String,
+    description: String,
+    flagged: Boolean,
+    category: String,
+    priority: String,
+    location: String,
+    active: Boolean
+)
